// src/proto-ts/frameworks/base/proto/src/apk_parser_config.ts
export interface ApkParsingConfig {
  skipUsesPermission: string[];
  flags: number;
}

export enum ApkParsingConfig_Flag {
  isolate_from_gmscore_and_finsky = 0,
  isolate_from_user_apps = 1,
  include_as_untrusted_app = 2,
  UNRECOGNIZED = -1,
}
// ...
export interface ApkParserConfig {
  /** permission name -> package name */
  permissions: { [key: string]: string };
  /** permission group name -> package name */
  permissionGroups: { [key: string]: string };
  /** content provider authority -> package name */
  contentProviderAuthorities: { [key: string]: string };
  /** packages from stock OS that are not allowed to be installed */
  nonInstallablePackages: string[];
  /** key is package name */
  parsingConfigs: { [key: string]: ApkParsingConfig };
  /**
   * packages from stock OS that are allowed to be installed if the certificate digest matches
   * package name -> cert SHA-256 digest
   */
  installablePackages: { [key: string]: string };
}
// ...
export const ApkParsingConfig: MessageFns<ApkParsingConfig> = {
  encode(message: ApkParsingConfig, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.skipUsesPermission) {
      writer.uint32(10).string(v!);
    }
    if (message.flags !== 0) {
      writer.uint32(16).int64(message.flags);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ApkParsingConfig {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseApkParsingConfig();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.skipUsesPermission.push(reader.string());
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.flags = longToNumber(reader.int64());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ApkParsingConfig {
    return {
      skipUsesPermission: globalThis.Array.isArray(object?.skipUsesPermission)
        ? object.skipUsesPermission.map((e: any) => globalThis.String(e))
        : globalThis.Array.isArray(object?.skip_uses_permission)
        ? object.skip_uses_permission.map((e: any) => globalThis.String(e))
        : [],
      flags: isSet(object.flags) ? globalThis.Number(object.flags) : 0,
    };
  },

  toJSON(message: ApkParsingConfig): unknown {
    const obj: any = {};
    if (message.skipUsesPermission?.length) {
      obj.skipUsesPermission = message.skipUsesPermission;
    }
    if (message.flags !== 0) {
      obj.flags = Math.round(message.flags);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ApkParsingConfig>, I>>(base?: I): ApkParsingConfig {
    return ApkParsingConfig.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ApkParsingConfig>, I>>(object: I): ApkParsingConfig {
    const message = createBaseApkParsingConfig();
    message.skipUsesPermission = object.skipUsesPermission?.map((e) => e) || [];
    message.flags = object.flags ?? 0;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}